Update README.md
While on the one hand, this app could seem to sensationalize the increasingly reckless and hostile behaviors manifesting on American roadways as the extension of hostilities between our increasingly militarized demographic factions,- which seem a prelude to all out, apocalyptic, class warfare. 

The 'idrivetooluicklyapp.app,' feature- while novel, is intended ironically- as a red herring. 
The red herring is necessary as a means of throwing members of the establishment- and their bots- off the scent of subversion so that those of us outsiders and members of the counter-culture can air our grievances, about the misdoings of our peers in the alt-right, the billionaire+ class, and all branches of govt, and hold them-- and one another-- to account. There is  an un-hyped and understated social media aspect of the platform where people with reasoned and intelligible viewpoints can post angsty diatribes, left-leaning political screeds, eyewitness accounts, and other personal philosophical and artistic statements about navigating the  consumerist trappings  that constrain us during this era of cultural collapse-